refactor of component structure
