+++ resources/js/Components/Inputs/InputLabel.vue
