+++ resources/js/Layouts/AuthenticatedLayout.vue
@@ -1,10 +1,10 @@
 <script setup>
 import { ref } from "vue";
 import ApplicationLogo from "@/Components/ApplicationLogo.vue";
-import Dropdown from "@/Components/Dropdown.vue";
-import DropdownLink from "@/Components/DropdownLink.vue";
-import NavLink from "@/Components/NavLink.vue";
-import ResponsiveNavLink from "@/Components/ResponsiveNavLink.vue";
+import Dropdown from "@/Components/Dropdowns/Dropdown.vue";
+import DropdownLink from "@/Components/Dropdowns/DropdownLink.vue";
+import NavLink from "@/Components/Links/NavLink.vue";
+import ResponsiveNavLink from "@/Components/Links/ResponsiveNavLink.vue";
 import { Link } from "@inertiajs/vue3";
 
 const showingNavigationDropdown = ref(false);
