+++ resources/js/Pages/Auth/Register.vue
@@ -1,9 +1,9 @@
 <script setup>
 import GuestLayout from "@/Layouts/GuestLayout.vue";
-import InputError from "@/Components/InputError.vue";
-import InputLabel from "@/Components/InputLabel.vue";
-import PrimaryButton from "@/Components/PrimaryButton.vue";
-import TextInput from "@/Components/TextInput.vue";
+import InputError from "@/Components/Inputs/InputError.vue";
+import InputLabel from "@/Components/Inputs/InputLabel.vue";
+import PrimaryButton from "@/Components/Buttons/PrimaryButton.vue";
+import TextInput from "@/Components/Inputs/TextInput.vue";
 import { Head, Link, useForm } from "@inertiajs/vue3";
 
 const form = useForm({
