+++ resources/js/Pages/Auth/VerifyEmail.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { computed } from "vue";
 import GuestLayout from "@/Layouts/GuestLayout.vue";
-import PrimaryButton from "@/Components/PrimaryButton.vue";
+import PrimaryButton from "@/Components/Buttons/PrimaryButton.vue";
 import { Head, Link, useForm } from "@inertiajs/vue3";
 
 const props = defineProps({
@@ -53,8 +53,8 @@ const verificationLinkSent = computed(
                     method="post"
                     as="button"
                     class="underline text-sm text-gray-600 hover:text-gray-900 rounded-md focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-                    >Log Out</Link
-                >
+                    >Log Out
+                </Link>
             </div>
         </form>
     </GuestLayout>
