+++ resources/js/Pages/Profile/Partials/DeleteUserForm.vue
@@ -1,10 +1,10 @@
 <script setup>
-import DangerButton from "@/Components/DangerButton.vue";
-import InputError from "@/Components/InputError.vue";
-import InputLabel from "@/Components/InputLabel.vue";
+import DangerButton from "@/Components/Buttons/DangerButton.vue";
+import InputError from "@/Components/Inputs/InputError.vue";
+import InputLabel from "@/Components/Inputs/InputLabel.vue";
 import Modal from "@/Components/Modal.vue";
-import SecondaryButton from "@/Components/SecondaryButton.vue";
-import TextInput from "@/Components/TextInput.vue";
+import SecondaryButton from "@/Components/Buttons/SecondaryButton.vue";
+import TextInput from "@/Components/Inputs/TextInput.vue";
 import { useForm } from "@inertiajs/vue3";
 import { nextTick, ref } from "vue";
