+++ resources/js/Pages/Profile/Partials/UpdatePasswordForm.vue
@@ -1,8 +1,8 @@
 <script setup>
-import InputError from "@/Components/InputError.vue";
-import InputLabel from "@/Components/InputLabel.vue";
-import PrimaryButton from "@/Components/PrimaryButton.vue";
-import TextInput from "@/Components/TextInput.vue";
+import InputError from "@/Components/Inputs/InputError.vue";
+import InputLabel from "@/Components/Inputs/InputLabel.vue";
+import PrimaryButton from "@/Components/Buttons/PrimaryButton.vue";
+import TextInput from "@/Components/Inputs/TextInput.vue";
 import { useForm } from "@inertiajs/vue3";
 import { ref } from "vue";
